[#114] Fix report by user functionality
Report by user did not take into account that user might have many channels.
So, I created a method that searches for user in DB and returns many instances
of the user type with different channels.
Then there was a problem with parsing that data, so i had to create a separate
type for type reportEntry which will hold a body or the entry when there are
many channels for one user
Then I changed the method to parse reportEntries into string with previous changes
After it I added new translation for output channel

Need to add tests that take into account new changes. Need to include several test cases
1. User has only one project and did not write standup
2. User have one project and wrote
standup.
3. User has many projects and did not write standup in any of them
4. User has many projects and wrote standups in some of them